# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Roster Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  inviteCode?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  manager: "manager",
  viewer: "viewer",
} as const;

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  role: UserProfileRole;
  mfaEnabled: boolean;
  createdAt: string;
  lastLogin?: string | null;
}

export interface LoginResponse {
  requiresMfa: boolean;
  mfaToken?: string;
  user?: UserProfile;
  message?: string;
}

export interface VerifyMfaRequest {
  mfaToken: string;
  code: string;
}

export interface AuthResponse {
  user: UserProfile;
  message?: string;
}

export interface MfaSetupResponse {
  secret: string;
  qrCodeUrl: string;
  backupCodes?: string[];
}

export interface ConfirmMfaRequest {
  code: string;
  secret: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  manager: "manager",
  viewer: "viewer",
} as const;

export interface CreateUserRequest {
  username: string;
  email: string;
  password: string;
  role: CreateUserRequestRole;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  manager: "manager",
  viewer: "viewer",
} as const;

export interface UpdateUserRequest {
  username?: string;
  email?: string;
  role?: UpdateUserRequestRole;
  password?: string;
}

export interface Rank {
  id: number;
  name: string;
  /** Higher number = higher authority */
  level: number;
  abbreviation: string;
  description?: string | null;
  memberCount?: number;
}

export interface CreateRankRequest {
  name: string;
  level: number;
  abbreviation: string;
  description?: string;
}

export interface UpdateRankRequest {
  name?: string;
  level?: number;
  abbreviation?: string;
  description?: string;
}

export interface Squad {
  id: number;
  name: string;
  description?: string | null;
  memberCount?: number;
}

export interface CreateSquadRequest {
  name: string;
  description?: string;
}

export interface UpdateSquadRequest {
  name?: string;
  description?: string;
}

export type RosterMemberStatus =
  (typeof RosterMemberStatus)[keyof typeof RosterMemberStatus];

export const RosterMemberStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface RosterMember {
  id: number;
  username: string;
  displayName?: string | null;
  rankId?: number | null;
  rankName?: string | null;
  rankAbbreviation?: string | null;
  squadId?: number | null;
  squadName?: string | null;
  status: RosterMemberStatus;
  notes?: string | null;
  joinedAt: string;
  updatedAt: string;
}

export type CreateMemberRequestStatus =
  (typeof CreateMemberRequestStatus)[keyof typeof CreateMemberRequestStatus];

export const CreateMemberRequestStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface CreateMemberRequest {
  username: string;
  displayName?: string;
  rankId?: number;
  squadId?: number;
  status?: CreateMemberRequestStatus;
  notes?: string;
}

export type UpdateMemberRequestStatus =
  (typeof UpdateMemberRequestStatus)[keyof typeof UpdateMemberRequestStatus];

export const UpdateMemberRequestStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface UpdateMemberRequest {
  username?: string;
  displayName?: string;
  rankId?: number | null;
  squadId?: number | null;
  status?: UpdateMemberRequestStatus;
  notes?: string;
}

export interface ActivityLog {
  id: number;
  userId?: number | null;
  userUsername?: string | null;
  action: string;
  entityType?: string | null;
  entityId?: number | null;
  details?: string | null;
  ipAddress?: string | null;
  createdAt: string;
}

export interface ActivityLogList {
  logs: ActivityLog[];
  total: number;
}

export interface RankCount {
  rankId: number;
  rankName: string;
  count: number;
}

export interface SquadCount {
  squadId: number;
  squadName: string;
  count: number;
}

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  totalSquads: number;
  totalRanks: number;
  recentActivity: ActivityLog[];
  membersByRank: RankCount[];
  membersBySquad: SquadCount[];
}

export type ListActivityLogsParams = {
  limit?: number;
  offset?: number;
};