# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Roster Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-mfa:
    post:
      operationId: verifyMfa
      tags: [auth]
      summary: Verify MFA code and complete login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyMfaRequest"
      responses:
        "200":
          description: MFA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/setup-mfa:
    post:
      operationId: setupMfa
      tags: [auth]
      summary: Setup MFA for the current user
      responses:
        "200":
          description: MFA setup info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaSetupResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/confirm-mfa:
    post:
      operationId: confirmMfa
      tags: [auth]
      summary: Confirm and enable MFA after setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmMfaRequest"
      responses:
        "200":
          description: MFA enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ranks:
    get:
      operationId: listRanks
      tags: [ranks]
      summary: List all rank levels
      responses:
        "200":
          description: List of ranks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rank"
    post:
      operationId: createRank
      tags: [ranks]
      summary: Create a rank level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRankRequest"
      responses:
        "201":
          description: Created rank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rank"
  /ranks/{id}:
    put:
      operationId: updateRank
      tags: [ranks]
      summary: Update a rank level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRankRequest"
      responses:
        "200":
          description: Updated rank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rank"
    delete:
      operationId: deleteRank
      tags: [ranks]
      summary: Delete a rank level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        inviteCode:
          type: string
      required: [username, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        requiresMfa:
          type: boolean
        mfaToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: string
      required: [requiresMfa]
    VerifyMfaRequest:
      type: object
      properties:
        mfaToken:
          type: string
        code:
          type: string
      required: [mfaToken, code]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: string
      required: [user]
    MfaSetupResponse:
      type: object
      properties:
        secret:
          type: string
        qrCodeUrl:
          type: string
        backupCodes:
          type: array
          items:
            type: string
      required: [secret, qrCodeUrl]
    ConfirmMfaRequest:
      type: object
      properties:
        code:
          type: string
        secret:
          type: string
      required: [code, secret]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
      required: [id, username, email, role, mfaEnabled, createdAt]
...
    Rank:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
          description: Higher number = higher authority
        abbreviation:
          type: string
        description:
          type: string
          nullable: true
        memberCount:
          type: integer
      required: [id, name, level, abbreviation]
    CreateRankRequest:
      type: object
      properties:
        name:
          type: string
        level:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
      required: [name, level, abbreviation]
    UpdateRankRequest:
      type: object
      properties:
        name:
          type: string
        level:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
...